+++ pkg/resource/deploy/source_eval_test.go
@@ -55,7 +55,7 @@ func fixedProgram(steps []RegisterResourceEvent) deploytest.ProgramFunc {
 		for _, s := range steps {
 			g := s.Goal()
 			urn, id, outs, err := resmon.RegisterResource(g.Type, string(g.Name), g.Custom, g.Parent, g.Protect,
-				g.Dependencies, g.Provider, g.Properties, g.PropertyDependencies, false, "")
+				g.Dependencies, g.Provider, g.Properties, g.PropertyDependencies, false, "", nil)
 			if err != nil {
 				return err
 			}
@@ -73,7 +73,7 @@ func newTestPluginContext(program deploytest.ProgramFunc) (*plugin.Context, erro
 	statusSink := cmdutil.Diag()
 	lang := deploytest.NewLanguageRuntime(program)
 	host := deploytest.NewPluginHost(sink, statusSink, lang)
-	return plugin.NewContext(sink, statusSink, host, nil, nil, "", nil, nil)
+	return plugin.NewContext(sink, statusSink, host, nil, "", nil, nil)
 }
 
 type testProviderSource struct {
@@ -143,16 +143,16 @@ func TestRegisterNoDefaultProviders(t *testing.T) {
 		// Register a component resource.
 		&testRegEvent{
 			goal: resource.NewGoal(componentURN.Type(), componentURN.Name(), false, resource.PropertyMap{}, "", false,
-				nil, "", []string{}, nil, false),
+				nil, "", []string{}, nil, false, nil),
 		},
 		// Register a couple resources using provider A.
 		&testRegEvent{
 			goal: resource.NewGoal("pkgA:index:typA", "res1", true, resource.PropertyMap{}, componentURN, false, nil,
-				providerARef.String(), []string{}, nil, false),
+				providerARef.String(), []string{}, nil, false, nil),
 		},
 		&testRegEvent{
 			goal: resource.NewGoal("pkgA:index:typA", "res2", true, resource.PropertyMap{}, componentURN, false, nil,
-				providerARef.String(), []string{}, nil, false),
+				providerARef.String(), []string{}, nil, false, nil),
 		},
 		// Register two more providers.
 		newProviderEvent("pkgA", "providerB", nil, ""),
@@ -160,11 +160,11 @@ func TestRegisterNoDefaultProviders(t *testing.T) {
 		// Register a few resources that use the new providers.
 		&testRegEvent{
 			goal: resource.NewGoal("pkgB:index:typB", "res3", true, resource.PropertyMap{}, "", false, nil,
-				providerBRef.String(), []string{}, nil, false),
+				providerBRef.String(), []string{}, nil, false, nil),
 		},
 		&testRegEvent{
 			goal: resource.NewGoal("pkgB:index:typC", "res4", true, resource.PropertyMap{}, "", false, nil,
-				providerCRef.String(), []string{}, nil, false),
+				providerCRef.String(), []string{}, nil, false, nil),
 		},
 	}
 
@@ -227,25 +227,25 @@ func TestRegisterDefaultProviders(t *testing.T) {
 		// Register a component resource.
 		&testRegEvent{
 			goal: resource.NewGoal(componentURN.Type(), componentURN.Name(), false, resource.PropertyMap{}, "", false,
-				nil, "", []string{}, nil, false),
+				nil, "", []string{}, nil, false, nil),
 		},
 		// Register a couple resources from package A.
 		&testRegEvent{
 			goal: resource.NewGoal("pkgA:m:typA", "res1", true, resource.PropertyMap{},
-				componentURN, false, nil, "", []string{}, nil, false),
+				componentURN, false, nil, "", []string{}, nil, false, nil),
 		},
 		&testRegEvent{
 			goal: resource.NewGoal("pkgA:m:typA", "res2", true, resource.PropertyMap{},
-				componentURN, false, nil, "", []string{}, nil, false),
+				componentURN, false, nil, "", []string{}, nil, false, nil),
 		},
 		// Register a few resources from other packages.
 		&testRegEvent{
 			goal: resource.NewGoal("pkgB:m:typB", "res3", true, resource.PropertyMap{}, "", false,
-				nil, "", []string{}, nil, false),
+				nil, "", []string{}, nil, false, nil),
 		},
 		&testRegEvent{
 			goal: resource.NewGoal("pkgB:m:typC", "res4", true, resource.PropertyMap{}, "", false,
-				nil, "", []string{}, nil, false),
+				nil, "", []string{}, nil, false, nil),
 		},
 	}
 
@@ -606,17 +606,17 @@ func TestRegisterResourceWithVersion(t *testing.T) {
 	program := func(_ plugin.RunInfo, resmon *deploytest.ResourceMonitor) error {
 		// Triggers pkgA, v0.18.1.
 		_, _, _, err := resmon.RegisterResource("pkgA:m:typA", "resA", true, "", false, nil, "",
-			resource.PropertyMap{}, nil, false, "0.18.1")
+			resource.PropertyMap{}, nil, false, "0.18.1", nil)
 		assert.NoError(t, err)
 
 		// Re-uses pkgA's already-instantiated provider.
 		_, _, _, err = resmon.RegisterResource("pkgA:m:typA", "resB", true, "", false, nil, "",
-			resource.PropertyMap{}, nil, false, "0.18.1")
+			resource.PropertyMap{}, nil, false, "0.18.1", nil)
 		assert.NoError(t, err)
 
 		// Triggers pkgA, v0.18.2
 		_, _, _, err = resmon.RegisterResource("pkgA:m:typA", "resB", true, "", false, nil, "",
-			resource.PropertyMap{}, nil, false, "0.18.2")
+			resource.PropertyMap{}, nil, false, "0.18.2", nil)
 		assert.NoError(t, err)
 		return nil
 	}