+++ sdk/python/lib/pulumi/runtime/invoke.py
@@ -50,8 +50,9 @@ def invoke(tok: str, props: Inputs, opts: InvokeOptions = None) -> Awaitable[Any
 
         monitor = get_monitor()
         inputs = await rpc.serialize_properties(props, {})
+        version = opts.version or ""
         log.debug(f"Invoking function prepared: tok={tok}")
-        req = provider_pb2.InvokeRequest(tok=tok, args=inputs, provider=provider_ref)
+        req = provider_pb2.InvokeRequest(tok=tok, args=inputs, provider=provider_ref, version=version)
 
         def do_invoke():
             try: