+++ sdk/python/lib/test/langhost/output_nested/test_output_nested.py
@@ -22,7 +22,8 @@ class OutputNestedTest(LanghostTest):
             expected_resource_count=3)
 
     def register_resource(self, _ctx, _dry_run, ty, name, resource,
-                          _dependencies, _parent, _custom, _protect, _provider, _property_deps, _delete_before_replace):
+                          _dependencies, _parent, _custom, _protect, _provider, _property_deps, _delete_before_replace,
+                          _ignore_changes, _version):
         nested_numbers = None
         if name == "testResource1":
             self.assertEqual(ty, "test:index:MyResource")