+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-## 0.17.8 (Unreleased)
+## 0.17.9 (Unreleased)
 
 ### Improvements
+
+## 0.17.8 (Released April 23, 2019)
+
+### Improvements
+
+- Add a new `ignoreChanges` option to resource options to allow specifying a list of properties to
+  ignore for purposes of updates or replacements.  [#2657](https://github.com/pulumi/pulumi/pull/2657)
+- Fix an engine bug that could lead to incorrect interpretation of the previous state of a resource leading to 
+  unexpected Update, Replace or Delete operations being scheduled. [#2650]https://github.com/pulumi/pulumi/issues/2650)
+- Build/push `pulumi/actions` container to [DockerHub](https://hub.docker.com/r/pulumi/actions) with new SDK releases [#2646](https://github.com/pulumi/pulumi/pull/2646)
 
 ## 0.17.7 (Released April 17, 2019)